+++ tests/test_NIRSpec.py
@@ -15,7 +15,7 @@ from eureka.S5_lightcurve_fitting import s5_fit as s5
 from tests.s3_reference import write_s3_reference_metadata
 
 
-def test_NIRSpec(capsys, keep_s3_output):
+def test_NIRSpec(capsys, keep_s2_output, keep_s3_output):
     with capsys.disabled():
         # is able to display any message without failing a test
         # useful to leave messages for future users who run the tests
@@ -96,7 +96,8 @@ def test_NIRSpec(capsys, keep_s3_output):
     assert np.array_equal(s5_meta.citations, s5_cites)
 
     # remove temporary files
-    os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}NIRSpec{os.sep}Stage2")
+    if not keep_s2_output:
+        os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}NIRSpec{os.sep}Stage2")
     if not keep_s3_output:
         os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}NIRSpec{os.sep}Stage3")
     os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}NIRSpec{os.sep}Stage4")
